Move Downloader.extract method to PathHelper class
