+++ src/libs/helpers/downloader.ts
@@ -2,8 +2,6 @@ import * as https from "https";
 import { IncomingMessage } from "http";
 import { RequestOptions } from "node:https";
 import * as fs from "fs";
-import * as path from "path";
-import jsZip from "jszip";
 
 export default class Downloader {
   private _accept: string = "*";
@@ -67,74 +65,6 @@ export default class Downloader {
         });
 
         resp.pipe(file);
-      });
-    });
-  }
-
-  /**
-   * Extract zip archive
-   * @param zipPath Path to zip file that should be uncompressed.
-   * @param destFolder Path to destination folder for uncompressed files.
-   * @param filesToExtract String array of file names, only files included in this array will be extracted. Leave empty to extract all files.
-   * @param deleteAfter If should delete zip file once finished extracting files.
-   */
-  public static extract(
-    zipPath: string,
-    destFolder: string,
-    filesToExtract: Array<string> = [],
-    deleteAfter: boolean = true
-  ) {
-    return new Promise((resolve, reject) => {
-      fs.readFile(zipPath, (err, data) => {
-        if (err) reject(err);
-
-        const zip = new jsZip();
-
-        zip.loadAsync(data).then((contents: jsZip) => {
-          const files = contents.files;
-
-          // Delete directories from object
-          for (const f in files) {
-            if (files[f].dir == true) delete files[f];
-          }
-
-          Object.keys(files).forEach(async (filename: string, i) => {
-            const filenameWithoutFolder = path.basename(filename);
-
-            // Write zip file to destination folder
-            const unzip = () => {
-              return new Promise((resolve) => {
-                zip
-                  .file(filename)!
-                  .async("nodebuffer")
-                  .then((content: any) => {
-                    fs.writeFile(path.join(destFolder, filenameWithoutFolder), content, () => {
-                      resolve("");
-                    });
-                  });
-              });
-            };
-
-            // If filesToExtract is empty just unzip all files
-            // If filesToExtract isn't empty, if it includes filename then unzip
-            if (filesToExtract.length == 0) await unzip();
-            else if (filesToExtract.includes(filenameWithoutFolder)) await unzip();
-
-            // Resolve if index (+1) equals amount of files since this means all files have been processed
-            if (i + 1 == Object.keys(files).length) {
-              // Delete zip file if asked to.
-              // Don't wait for file to delete before resolving the promise,
-              // theres no reason to make the user wait for it to finish deleting the zip.
-              if (deleteAfter) {
-                fs.unlink(zipPath, (err) => {
-                  if (err) throw err;
-                });
-              }
-
-              resolve("");
-            }
-          });
-        });
       });
     });
   }
